show experiments in the version ui
